Split up into individual declaration files
