+++ Sources/Ignite/Framework/CoreAttributes.swift
@@ -6,43 +6,6 @@
 //
 
 import Foundation
-import OrderedCollections
-
-// A typealias that allows us to use `OrderedSet` without importing OrderedCollections
-public typealias OrderedSet<Element: Hashable> = OrderedCollections.OrderedSet<Element>
-
-// A typealias that allows us to use `OrderedDictionary` without importing OrderedCollections
-public typealias OrderedDictionary<Key: Hashable, Value> = OrderedCollections.OrderedDictionary<Key, Value>
-
-// A typealias that allows us to use `UUID` without importing Foundation
-public typealias UUID = Foundation.UUID
-
-// A typealias that allows us to use `URL` without importing Foundation
-public typealias URL = Foundation.URL
-
-// A typealias that allows us to use `Data` without importing Foundation
-public typealias Data = Foundation.Data
-
-// A typealias that allows us to use `Date` without importing Foundation
-public typealias Date = Foundation.Date
-
-/// A publishing-time registration required by rendered attributes.
-enum PublishingRegistration: Hashable, Equatable, Sendable {
-    case fontFamily(Font)
-    case responsiveFont(ResponsiveValues<LengthUnit>)
-    case responsiveVisibility(ResponsiveValues<Bool>)
-
-    func apply(to context: PublishingContext) {
-        switch self {
-        case .fontFamily(let font):
-            context.cssManager.registerFontFamily(font)
-        case .responsiveFont(let values):
-            _ = context.cssManager.registerFont(values)
-        case .responsiveVisibility(let values):
-            _ = context.cssManager.registerStyles(values)
-        }
-    }
-}
 
 /// A handful of attributes that all HTML types must support, either for
 /// rendering or for publishing purposes.
