+++ Sources/Ignite/Framework/StringInterpolation+CoreAttributes.swift
@@ -1,0 +1,7 @@
+//
+// StringInterpolation+CoreAttributes.swift
+// Ignite
+// https://www.github.com/twostraws/Ignite
+// See LICENSE for license information.
+//
+
